requirements.txt
alembic==0.9.9
arrow==0.12.1
Django==2.0.5
django-cors-headers==2.2.0
django-extensions==2.0.7
django-model-utils==3.1.2
future==0.16.0
idna==2.6
isodate==0.6.0
Jinja2==2.10
jinja2-time==0.2.0
Mako==1.0.7
MarkupSafe==1.0
poyo==0.4.1
psycopg2==2.7.4
psycopg2-binary==2.7.4
pyparsing==2.2.0
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2018.4
rdflib==4.2.2
rdflib-django==0.3
rdflib-sqlalchemy==0.3.8
six==1.11.0
SPARQLWrapper==1.8.2
SQLAlchemy==1.2.8
SuRF==1.1.9
Unipath==1.1
